fix an exploding memory view change list, causing hangs
This patch fixes Issues #3 and #10, and other similar problems.
Essentially Felix broke in commits
- 256f299be50229bf2da0b4d4cb985489037b303b and
- e3f68aa88e76fce575abb4b42c20aa38e2cbf35c
how the `Environment` keeps track of changes which need to later be
painted by the GUI in red (using HTML). As the `MyByte.equals` and
`MyByte.hash` were changed, the checks for adding HTML were querying
MyByte Values, instead of the MyByte locations. For example, if a 0x00
byte lands in the initial list of changes, all 0x00 bytes would be
painted...

The fix is simple: drop the old change list and use only the newer &
better `changedCells` hash map that Felix introduced.

I caught this problem after regression testing a previous simulator
version (1.10).

The painting with HTML is still not optimal. At least it now is somehow
kept at bay.

Signed-off-by: Vasil Sarafov <contact@sarafov.net>